AA+ v1.71 -> 1.73

+++ aaplus-v1.73/AAEaster.h
@@ -52,7 +52,7 @@ class AAPLUS_EXT_CLASS CAAEaster
 {
 public:
 //Static methods
-  static CAAEasterDetails Calculate(int nYear, bool GregorianCalendar);
+  static CAAEasterDetails Calculate(long nYear, bool GregorianCalendar);
 };
